chore(ui): extract moveDocuments so drop and the move button share one path
